the last update

+++ src/main/java/com/echo/acknowledgehub/bean/SystemDataBean.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Component;
 public class SystemDataBean {
     private volatile boolean checkPasswordEveryTime;
     private String defaultPassword;
-    private int sessionExpireTime;
+    private long sessionExpireTime;
     private boolean restSystem;
 
     //Add your system data field(Important data) here. Only primitive type allow here.
